Remove libgcc_s.so from initramfs
libgcc_s.so is pulled into intramfs because anything that calls
pthread_cancel() will trigger a dlopen() of libgcc_s.so. If it's
not present, the caller will exit with an error message.

libgcc_s.so is GPLv3 licensed with a quirky exception clause that
appears to only grant exception to the viral clause of the GPL.
The anti-tivoization clause may still be in effect.

A quick audit indicates that we don't have any pthread_cancel
callers in our startup, so we're removing libgcc_s.so.

https://phabricator.endlessm.com/T30878